+++ test/index.js
@@ -2,5 +2,6 @@
 
 exports["test reduce"] = require("./reduce")
 exports["test reducible"] = require("./reducible")
+exports["test errors"] = require("./error")
 
 require("test").run(exports)
